Consolidate dashboard filters reproductions
