--- e2e/test/scenarios/dashboard-filters/reproductions/22482-round-relative-ranges.cy.spec.js
@@ -1,53 +1,0 @@
-import moment from "moment-timezone"; // eslint-disable-line no-restricted-imports -- deprecated usage
-
-import { ORDERS_DASHBOARD_ID } from "e2e/support/cypress_sample_instance_data";
-import {
-  restore,
-  popover,
-  filterWidget,
-  editDashboard,
-  saveDashboard,
-  setFilter,
-  visitDashboard,
-} from "e2e/support/helpers";
-
-describe("issue 22482", () => {
-  beforeEach(() => {
-    restore();
-    cy.signInAsAdmin();
-
-    visitDashboard(ORDERS_DASHBOARD_ID);
-
-    editDashboard();
-    setFilter("Time", "All Options");
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Select…").click();
-    popover().contains("Created At").eq(0).click();
-
-    saveDashboard();
-
-    filterWidget().click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("Relative dates...").click();
-  });
-
-  it("should round relative date range (metabase#22482)", () => {
-    cy.findByTestId("relative-datetime-value").clear().type(15);
-    cy.findByTestId("relative-datetime-unit").click();
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText("months").click();
-
-    const expectedRange = getFormattedRange(
-      moment().startOf("month").add(-15, "month"),
-      moment().add(-1, "month").endOf("month"),
-    );
-
-    // eslint-disable-next-line no-unscoped-text-selectors -- deprecated usage
-    cy.findByText(expectedRange);
-  });
-});
-
-function getFormattedRange(start, end) {
-  return `${start.format("MMM D, YYYY")} - ${end.format("MMM D, YYYY")}`;
-}
